Webapp refactoring, web page
* Added a simple index.html page with templating to display the latest
  temperature from the latest "session"
* Refactored the webapp code a bit
