+++ Cargo.toml
@@ -15,8 +15,9 @@ bytes = "1"
 chrono = "0.4.31"
 env_logger = "0.10.1"
 futures = "0.3.29"
-hyper = { version = "1", features = ["full"] }
+handlebars = "5"
 http-body-util = "0.1"
+hyper = { version = "1", features = ["full"] }
 hyper-util = { version = "0.1", features = ["full"] }
 log = "0.4.20"
 tokio = { version = "1", features = ["full"] }
