+++ src/bin/webapp/Dockerfile
@@ -1,0 +1,3 @@
+FROM amazonlinux:latest
+
+yum update
